Improve unit tests by moving common code into helper.
